move EventEmitter to utils
